+++ src/components/CodeBlock/CodeBlock.tsx
@@ -1,48 +1,69 @@
 import React from "react";
-// @ts-ignore
-(typeof global !== "undefined" ? global : (window as any)).Prism = Prism;
-import PrismReactRenderer, { defaultProps, Prism, Language } from "prism-react-renderer";
+import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter';
+import bash from 'react-syntax-highlighter/dist/esm/languages/prism/bash';
+import csharp from 'react-syntax-highlighter/dist/esm/languages/prism/csharp';
+import css from 'react-syntax-highlighter/dist/esm/languages/prism/css';
+import diff from 'react-syntax-highlighter/dist/esm/languages/prism/diff';
+import django from 'react-syntax-highlighter/dist/esm/languages/prism/django';
+import go from 'react-syntax-highlighter/dist/esm/languages/prism/go';
+import http from 'react-syntax-highlighter/dist/esm/languages/prism/http';
+import java from 'react-syntax-highlighter/dist/esm/languages/prism/java';
+import javascript from 'react-syntax-highlighter/dist/esm/languages/prism/javascript';
+import json from 'react-syntax-highlighter/dist/esm/languages/prism/json';
+import jsx from 'react-syntax-highlighter/dist/esm/languages/prism/jsx';
+import makefile from 'react-syntax-highlighter/dist/esm/languages/prism/makefile';
+import markup from 'react-syntax-highlighter/dist/esm/languages/prism/markup';
+import php from 'react-syntax-highlighter/dist/esm/languages/prism/php';
+import python from 'react-syntax-highlighter/dist/esm/languages/prism/python';
+import ruby from 'react-syntax-highlighter/dist/esm/languages/prism/ruby';
+import scala from 'react-syntax-highlighter/dist/esm/languages/prism/scala';
+import typescript from 'react-syntax-highlighter/dist/esm/languages/prism/typescript';
+import tsx from 'react-syntax-highlighter/dist/esm/languages/prism/tsx';
+import prism from 'react-syntax-highlighter/dist/esm/styles/prism/prism';
 
-require("prismjs/components/prism-java");
-require("prismjs/components/prism-csharp");
-require("prismjs/components/prism-scala");
-require("prismjs/components/prism-ruby");
-require("prismjs/components/prism-graphql");
-require("prismjs/components/prism-http");
+SyntaxHighlighter.registerLanguage('bash', bash);
+SyntaxHighlighter.registerLanguage('csharp', csharp);
+SyntaxHighlighter.registerLanguage('css', css);
+SyntaxHighlighter.registerLanguage('diff', diff);
+SyntaxHighlighter.registerLanguage('django', django);
+SyntaxHighlighter.registerLanguage('go', go);
+SyntaxHighlighter.registerLanguage('http', http);
+SyntaxHighlighter.registerLanguage('java', java);
+SyntaxHighlighter.registerLanguage('javascript', javascript);
+SyntaxHighlighter.registerLanguage('json', json);
+SyntaxHighlighter.registerLanguage('jsx', jsx);
+SyntaxHighlighter.registerLanguage('makefile', makefile);
+SyntaxHighlighter.registerLanguage('markup', markup);
+SyntaxHighlighter.registerLanguage('php', php);
+SyntaxHighlighter.registerLanguage('python', python);
+SyntaxHighlighter.registerLanguage('ruby', ruby);
+SyntaxHighlighter.registerLanguage('scala', scala);
+SyntaxHighlighter.registerLanguage('typescript', typescript);
+SyntaxHighlighter.registerLanguage('tsx', tsx);
 
 import { CommonAndHTMLProps } from "../Element/constants";
 
 // prettier-ignore
 export interface CodeBlockCustomProps {
     source   ? : object | string;
-    language ? : Language;
+    language ? : string;
 }
 
 export type CodeBlockElementType = HTMLPreElement;
 export type CodeBlockProps = CommonAndHTMLProps<CodeBlockElementType> & CodeBlockCustomProps;
 
-export const CodeBlock = React.forwardRef(
+
+export const CodeBlock = /*#__PURE__*/React.forwardRef(
     ({ source, ...props }: CodeBlockProps, ref: React.Ref<CodeBlockElementType>) => {
         return (
-            <PrismReactRenderer
-                {...defaultProps}
-                code={typeof source === "object" ? JSON.stringify(source, null, 2) : source}
-                language="json"
+            <SyntaxHighlighter 
+                language="json" 
                 {...props}
-                theme={undefined}
+                style={prism}
+                // useInlineStyles={false}
             >
-                {({ className, style, tokens, getLineProps, getTokenProps }) => (
-                    <pre ref={ref} className={className} style={style}>
-                        {tokens.map((line, i) => (
-                            <div {...getLineProps({ line, key: i })}>
-                                {line.map((token, key) => (
-                                    <span {...getTokenProps({ token, key })} />
-                                ))}
-                            </div>
-                        ))}
-                    </pre>
-                )}
-            </PrismReactRenderer>
+                {source}
+            </SyntaxHighlighter>
         );
     }
 );